==> 5. hktvmall_MergingDatasets.ipynb <==
import pandas as pd
#Load dataframes
df_hktvmall_all=pd.read_csv('hktvmall_skincare_females_and_males_cleaned.csv',index_col=0)
df_hktvmall_male=pd.read_csv('hktvmall_skincare_males_cleaned.csv',index_col=0)

#Rename Columns
df_hktvmall_all=df_hktvmall_all.rename(columns={"Discount price": "Discount Price"})
df_hktvmall_male=df_hktvmall_male.rename(columns={"Raw Product": "Raw Info", "Discount price": "Discount Price"})

#Create New Column - "Male_only" for Male Dataset
df_hktvmall_male["Male_only"]=df_hktvmall_male.apply(lambda _: '', axis=1)
df_hktvmall_male["Male_only"]=df_hktvmall_male["Male_only"].apply(lambda x: "Male")

#Combine Dataframe
df_combo=df_hktvmall_all.append(df_hktvmall_male)

#Handle nan values in "Male_only" column
df_combo["Male_only"]=df_combo["Male_only"].fillna("Female")

#Locate Duplicate
df_duplicate=df_combo.duplicated(subset=['Discount Price', 'Original Price', 'Product Brand',
       'Product Category Code', 'Product Name', 'Promo Tag', 'Rating',
       'Raw Info', 'Review', 'Sales', 'Vendor', 'Discount', 'Price Range',
       'Product Category', 'Revenue'], keep="first")

#Create New Column with Duplicate Marks
df_combo["Duplicate"]=df_duplicate

#Drop Duplicates
df_combo=df_combo.drop_duplicates(subset=['Discount Price', 'Original Price', 'Product Brand',
       'Product Category Code', 'Product Name', 'Promo Tag', 'Rating',
       'Raw Info', 'Review', 'Sales', 'Vendor', 'Discount', 'Price Range',
       'Product Category', 'Revenue'], keep='last')

#Create "Male_only" Column
df_combo["Duplicate"].value_counts()

#Save to csv
df_combo.to_csv("hktvmall_combine.csv")
